bugfixing hdf5 update

+++ spatial_access/src/userDataContainer.h
@@ -7,8 +7,8 @@ class userDataPoint
 public:
     int networkNodeId;
     unsigned int loc;
-    int lastMileDistance;
-    userDataPoint(int networkNodeId, unsigned int loc, int lastMileDistance);
+    unsigned short int lastMileDistance;
+    userDataPoint(unsigned int networkNodeId, unsigned int loc, unsigned short int lastMileDistance);
 };
 
 class userDataTract
@@ -16,7 +16,7 @@ class userDataTract
 public:
     int networkNodeId;
     std::vector<userDataPoint> data;
-    userDataTract(int networkNodeId);
+    userDataTract(unsigned int networkNodeId);
     void addPoint(userDataPoint userData);
     void print() const;
     const std::vector<userDataPoint>& retrieveDataPoints() const;
@@ -27,16 +27,16 @@ class userDataContainer
 {
 public:
     bool isSymmetric;
-    std::unordered_map<int, userDataTract> data;
-    std::vector<int> allNetworkNodeIds;
+    std::unordered_map<unsigned int, userDataTract> data;
+    std::vector<unsigned int> allNetworkNodeIds;
     std::vector<unsigned int> ids;
-    std::vector<int> uniqueNetworkNodeIds;
+    std::vector<unsigned int> uniqueNetworkNodeIds;
     userDataContainer();
-    void addPoint(int networkNodeId, unsigned int loc, int lastMileDistance);
-    bool containsTract(int networkNodeId) const;
-    const userDataTract& retrieveTract(int networkNodeId) const;
-    const std::vector<unsigned  int>& retrieveAllUserDataIds() const;
-    const std::vector<int>& retrieveAllNetworkNodeIds() const;
-    const std::vector<int>& retrieveUniqueNetworkNodeIds() const;
+    void addPoint(unsigned int networkNodeId, unsigned int loc, unsigned short int lastMileDistance);
+    bool containsTract(unsigned int networkNodeId) const;
+    const userDataTract& retrieveTract(unsigned int networkNodeId) const;
+    const std::vector<unsigned int>& retrieveAllUserDataIds() const;
+    const std::vector<unsigned int>& retrieveAllNetworkNodeIds() const;
+    const std::vector<unsigned int>& retrieveUniqueNetworkNodeIds() const;
     void print() const;
 };
